Refacator code in modules
