+++ src/modules/notifications.js
@@ -1,0 +1,11 @@
+const showMessage = (meassage) => {
+  const postMessage = document.getElementById('post-message');
+  postMessage.innerHTML = meassage;
+  postMessage.classList.remove('hidden');
+  setTimeout(() => {
+    postMessage.classList.add('hidden');
+    postMessage.innerHTML = '';
+  }, 3000);
+};
+
+export default showMessage;
